translate graphController to ts
